Put the rest of the memo classes into their own files.
